♻️ (#107): split heading logic: extractHeadings
