+++ src/extensions/RichText.js
@@ -38,7 +38,7 @@ import HardBreak from './HardBreak.js'
 import KeepSyntax from './KeepSyntax.js'
 import Table from './../nodes/Table.js'
 import Image from './../nodes/Image.js'
-import Heading from './../nodes/Heading.js'
+import Heading from '../nodes/Heading/index.js'
 import BulletList from './../nodes/BulletList.js'
 import TaskList from './../nodes/TaskList.js'
 import TaskItem from './../nodes/TaskItem.js'
